etl/raw_to_mongo/src/pyproject.toml
[tool.poetry]
name = "Парсер турнирных игр кс:го"
version = "0.1.0"
description = "Регулярно проверяет наличие новых игр у поставщика и складывает их в БД."
authors = ["Билибин Сергей"]
readme = "README.md"

[tool.poetry.dependencies]
python=">=3.10"
python-dotenv="==1.0.0"
pandas="==2.0.3"
numpy="==1.25.2"
pydantic-settings="==2.0.2"
pymongo="==4.4.1"
tqdm="==4.66.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"